requirements.txt
opencv-python
Pillow
mediapipe
pywin32
PyYAML
